Move AuthenticationFeatureConfig to a separate file
